model-training/pyproject.toml
[project]
name = "model-training"
version = "0.1.0"
description = "Model training pipeline for flight traffic forecasting with MLflow"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "polars>=1.0.0",
    "boto3>=1.34.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "scikit-learn>=1.4.0",
    "mlflow>=2.10.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
